// src/app/habits/models/streak.ts
export interface Streak {
  count: number;
  startDate: Date;
  endDate: Date;
}
// export class Streak implements StreakInterface{
//     constructor(
//         public count: number,
//         public startDate: Date,
//         public endDate: Date
//         ){}

//     toString(){
//         return this.count +
//     }

// }